+++ sdk/resourcemanager/security/armsecurity/models.go
@@ -52,6 +52,61 @@ type AADSolutionProperties struct {
 	Workspace *ConnectedWorkspace `json:"workspace,omitempty"`
 }
 
+// APICollectionClientGetOptions contains the optional parameters for the APICollectionClient.Get method.
+type APICollectionClientGetOptions struct {
+	// placeholder for future optional parameters
+}
+
+// APICollectionClientListOptions contains the optional parameters for the APICollectionClient.List method.
+type APICollectionClientListOptions struct {
+	// placeholder for future optional parameters
+}
+
+// APICollectionOffboardingClientDeleteOptions contains the optional parameters for the APICollectionOffboardingClient.Delete
+// method.
+type APICollectionOffboardingClientDeleteOptions struct {
+	// placeholder for future optional parameters
+}
+
+// APICollectionOnboardingClientCreateOptions contains the optional parameters for the APICollectionOnboardingClient.Create
+// method.
+type APICollectionOnboardingClientCreateOptions struct {
+	// placeholder for future optional parameters
+}
+
+// APICollectionProperties - Describes the properties of an API collection.
+type APICollectionProperties struct {
+	// Additional data regarding the API collection.
+	AdditionalData map[string]*string `json:"additionalData,omitempty"`
+
+	// The display name of the Azure API Management API.
+	DisplayName *string `json:"displayName,omitempty"`
+}
+
+// APICollectionResponse - An API collection as represented by Defender for APIs.
+type APICollectionResponse struct {
+	// Describes the properties of an API collection.
+	Properties *APICollectionProperties `json:"properties,omitempty"`
+
+	// READ-ONLY; Resource Id
+	ID *string `json:"id,omitempty" azure:"ro"`
+
+	// READ-ONLY; Resource name
+	Name *string `json:"name,omitempty" azure:"ro"`
+
+	// READ-ONLY; Resource type
+	Type *string `json:"type,omitempty" azure:"ro"`
+}
+
+// APICollectionResponseList - Page of a list of API collections as represented by Defender for APIs.
+type APICollectionResponseList struct {
+	// READ-ONLY; The URI to fetch the next page.
+	NextLink *string `json:"nextLink,omitempty" azure:"ro"`
+
+	// READ-ONLY; API collections in this page.
+	Value []*APICollectionResponse `json:"value,omitempty" azure:"ro"`
+}
+
 // AccountConnectorsClientCreateOrUpdateOptions contains the optional parameters for the AccountConnectorsClient.CreateOrUpdate
 // method.
 type AccountConnectorsClientCreateOrUpdateOptions struct {
@@ -2372,35 +2427,6 @@ type ConnectorApplicationsClientListOptions struct {
 	// placeholder for future optional parameters
 }
 
-// ConnectorGovernanceRuleClientListOptions contains the optional parameters for the ConnectorGovernanceRuleClient.List method.
-type ConnectorGovernanceRuleClientListOptions struct {
-	// placeholder for future optional parameters
-}
-
-// ConnectorGovernanceRulesClientCreateOrUpdateOptions contains the optional parameters for the ConnectorGovernanceRulesClient.CreateOrUpdate
-// method.
-type ConnectorGovernanceRulesClientCreateOrUpdateOptions struct {
-	// placeholder for future optional parameters
-}
-
-// ConnectorGovernanceRulesClientDeleteOptions contains the optional parameters for the ConnectorGovernanceRulesClient.Delete
-// method.
-type ConnectorGovernanceRulesClientDeleteOptions struct {
-	// placeholder for future optional parameters
-}
-
-// ConnectorGovernanceRulesClientGetOptions contains the optional parameters for the ConnectorGovernanceRulesClient.Get method.
-type ConnectorGovernanceRulesClientGetOptions struct {
-	// placeholder for future optional parameters
-}
-
-// ConnectorGovernanceRulesExecuteStatusClientBeginGetOptions contains the optional parameters for the ConnectorGovernanceRulesExecuteStatusClient.BeginGet
-// method.
-type ConnectorGovernanceRulesExecuteStatusClientBeginGetOptions struct {
-	// Resumes the LRO from the provided token.
-	ResumeToken string
-}
-
 // ConnectorProperties - A set of properties that defines the security connector configuration.
 type ConnectorProperties struct {
 	// The security connector environment data.
@@ -3703,16 +3729,35 @@ type ErrorAdditionalInfo struct {
 	Type *string `json:"type,omitempty" azure:"ro"`
 }
 
+// ErrorDetail - The error detail.
+type ErrorDetail struct {
+	// READ-ONLY; The error additional info.
+	AdditionalInfo []*ErrorAdditionalInfo `json:"additionalInfo,omitempty" azure:"ro"`
+
+	// READ-ONLY; The error code.
+	Code *string `json:"code,omitempty" azure:"ro"`
+
+	// READ-ONLY; The error details.
+	Details []*ErrorDetail `json:"details,omitempty" azure:"ro"`
+
+	// READ-ONLY; The error message.
+	Message *string `json:"message,omitempty" azure:"ro"`
+
+	// READ-ONLY; The error target.
+	Target *string `json:"target,omitempty" azure:"ro"`
+}
+
+// ErrorResponse - Common error response for all Azure Resource Manager APIs to return error details for failed operations.
+// (This also follows the OData error response format.).
+type ErrorResponse struct {
+	// The error object.
+	Error *ErrorDetail `json:"error,omitempty"`
+}
+
 // ExecuteGovernanceRuleParams - Governance rule execution parameters
 type ExecuteGovernanceRuleParams struct {
 	// Describe if governance rule should be override
 	Override *bool `json:"override,omitempty"`
-}
-
-// ExecuteRuleStatus - Execute status of Security GovernanceRule over a given scope
-type ExecuteRuleStatus struct {
-	// READ-ONLY; Unique key for the execution of GovernanceRule
-	OperationID *string `json:"operationId,omitempty" azure:"ro"`
 }
 
 // ExternalSecuritySolution - Represents a security solution external to Microsoft Defender for Cloud which sends information
@@ -4053,9 +4098,9 @@ func (g *GithubScopeEnvironmentData) GetEnvironmentData() *EnvironmentData {
 	}
 }
 
-// GovernanceAssignment - Security GovernanceAssignment over a given scope
+// GovernanceAssignment - Governance assignment over a given scope
 type GovernanceAssignment struct {
-	// Properties of a security governanceAssignment
+	// The properties of a governance assignment
 	Properties *GovernanceAssignmentProperties `json:"properties,omitempty"`
 
 	// READ-ONLY; Resource Id
@@ -4068,15 +4113,15 @@ type GovernanceAssignment struct {
 	Type *string `json:"type,omitempty" azure:"ro"`
 }
 
-// GovernanceAssignmentAdditionalData - Describe the additional data of GovernanceAssignment - optional
+// GovernanceAssignmentAdditionalData - Describe the additional data of governance assignment - optional
 type GovernanceAssignmentAdditionalData struct {
-	// Ticket link associated with this GovernanceAssignment - for example: https://snow.com
+	// Ticket link associated with this governance assignment - for example: https://snow.com
 	TicketLink *string `json:"ticketLink,omitempty"`
 
-	// Ticket number associated with this GovernanceAssignment
+	// Ticket number associated with this governance assignment
 	TicketNumber *int32 `json:"ticketNumber,omitempty"`
 
-	// The ticket status associated with this GovernanceAssignment - for example: Active
+	// The ticket status associated with this governance assignment - for example: Active
 	TicketStatus *string `json:"ticketStatus,omitempty"`
 }
 
@@ -4122,7 +4167,7 @@ type GovernanceAssignmentsClientListOptions struct {
 	// placeholder for future optional parameters
 }
 
-// GovernanceAssignmentsList - Page of a security governance assignments list
+// GovernanceAssignmentsList - Page of a governance assignments list
 type GovernanceAssignmentsList struct {
 	// READ-ONLY; The URI to fetch the next page
 	NextLink *string `json:"nextLink,omitempty" azure:"ro"`
@@ -4140,9 +4185,9 @@ type GovernanceEmailNotification struct {
 	DisableOwnerEmailNotification *bool `json:"disableOwnerEmailNotification,omitempty"`
 }
 
-// GovernanceRule - Security GovernanceRule over a given scope
+// GovernanceRule - Governance rule over a given scope
 type GovernanceRule struct {
-	// Properties of a security governanceRule
+	// Properties of a governance rule
 	Properties *GovernanceRuleProperties `json:"properties,omitempty"`
 
 	// READ-ONLY; Resource Id
@@ -4155,27 +4200,37 @@ type GovernanceRule struct {
 	Type *string `json:"type,omitempty" azure:"ro"`
 }
 
-// GovernanceRuleClientListOptions contains the optional parameters for the GovernanceRuleClient.List method.
-type GovernanceRuleClientListOptions struct {
-	// placeholder for future optional parameters
-}
-
-// GovernanceRuleEmailNotification - The governance email weekly notification configuration.
+// GovernanceRuleEmailNotification - The governance email weekly notification configuration
 type GovernanceRuleEmailNotification struct {
-	// Defines whether manager email notifications are disabled.
+	// Defines whether manager email notifications are disabled
 	DisableManagerEmailNotification *bool `json:"disableManagerEmailNotification,omitempty"`
 
-	// Defines whether owner email notifications are disabled.
+	// Defines whether owner email notifications are disabled
 	DisableOwnerEmailNotification *bool `json:"disableOwnerEmailNotification,omitempty"`
 }
 
-// GovernanceRuleList - Page of a security governanceRules list
+// GovernanceRuleList - Page of a governance rules list
 type GovernanceRuleList struct {
 	// READ-ONLY; The URI to fetch the next page
 	NextLink *string `json:"nextLink,omitempty" azure:"ro"`
 
-	// READ-ONLY; Collection of governanceRules in this page
+	// READ-ONLY; Collection of governance rules in this page
 	Value []*GovernanceRule `json:"value,omitempty" azure:"ro"`
+}
+
+// GovernanceRuleMetadata - The governance rule metadata
+type GovernanceRuleMetadata struct {
+	// READ-ONLY; Governance rule Created by object id (GUID)
+	CreatedBy *string `json:"createdBy,omitempty" azure:"ro"`
+
+	// READ-ONLY; Governance rule creation date
+	CreatedOn *time.Time `json:"createdOn,omitempty" azure:"ro"`
+
+	// READ-ONLY; Governance rule last updated by object id (GUID)
+	UpdatedBy *string `json:"updatedBy,omitempty" azure:"ro"`
+
+	// READ-ONLY; Governance rule last update date
+	UpdatedOn *time.Time `json:"updatedOn,omitempty" azure:"ro"`
 }
 
 // GovernanceRuleOwnerSource - Describe the owner source of governance rule
@@ -4187,19 +4242,19 @@ type GovernanceRuleOwnerSource struct {
 	Value *string `json:"value,omitempty"`
 }
 
-// GovernanceRuleProperties - Describes properties of an governanceRule
+// GovernanceRuleProperties - Describes properties of an governance rule
 type GovernanceRuleProperties struct {
 	// REQUIRED; The governance rule conditionSets - see examples
 	ConditionSets []interface{} `json:"conditionSets,omitempty"`
 
-	// REQUIRED; display name of the governanceRule
+	// REQUIRED; Display name of the governance rule
 	DisplayName *string `json:"displayName,omitempty"`
 
-	// REQUIRED; The Owner source for the governance rule - e.g. Manually by user@contoso.com - see example
+	// REQUIRED; The owner source for the governance rule - e.g. Manually by user@contoso.com - see example
 	OwnerSource *GovernanceRuleOwnerSource `json:"ownerSource,omitempty"`
 
-	// REQUIRED; The governance rule priority, priority to the lower number. Rules with the same priority on the same subscription
-	// will not be allowed
+	// REQUIRED; The governance rule priority, priority to the lower number. Rules with the same priority on the same scope will
+	// not be allowed
 	RulePriority *int32 `json:"rulePriority,omitempty"`
 
 	// REQUIRED; The rule type of the governance rule, defines the source of the rule e.g. Integrated
@@ -4208,11 +4263,17 @@ type GovernanceRuleProperties struct {
 	// REQUIRED; The governance rule source, what the rule affects, e.g. Assessments
 	SourceResourceType *GovernanceRuleSourceResourceType `json:"sourceResourceType,omitempty"`
 
-	// description of the governanceRule
+	// Description of the governance rule
 	Description *string `json:"description,omitempty"`
+
+	// Excluded scopes, filter out the descendants of the scope (on management scopes)
+	ExcludedScopes []*string `json:"excludedScopes,omitempty"`
 
 	// The email notifications settings for the governance rule, states whether to disable notifications for mangers and owners
 	GovernanceEmailNotification *GovernanceRuleEmailNotification `json:"governanceEmailNotification,omitempty"`
+
+	// Defines whether the rule is management scope rule (master connector as a single scope or management scope)
+	IncludeMemberScopes *bool `json:"includeMemberScopes,omitempty"`
 
 	// Defines whether the rule is active/inactive
 	IsDisabled *bool `json:"isDisabled,omitempty"`
@@ -4220,24 +4281,26 @@ type GovernanceRuleProperties struct {
 	// Defines whether there is a grace period on the governance rule
 	IsGracePeriod *bool `json:"isGracePeriod,omitempty"`
 
+	// The governance rule metadata
+	Metadata *GovernanceRuleMetadata `json:"metadata,omitempty"`
+
 	// Governance rule remediation timeframe - this is the time that will affect on the grace-period duration e.g. 7.00:00:00
 	// - means 7 days
 	RemediationTimeframe *string `json:"remediationTimeframe,omitempty"`
+
+	// READ-ONLY; The tenantId (GUID)
+	TenantID *string `json:"tenantId,omitempty" azure:"ro"`
 }
 
-// GovernanceRulesClientBeginRuleIDExecuteSingleSecurityConnectorOptions contains the optional parameters for the GovernanceRulesClient.BeginRuleIDExecuteSingleSecurityConnector
-// method.
-type GovernanceRulesClientBeginRuleIDExecuteSingleSecurityConnectorOptions struct {
-	// GovernanceRule over a subscription scope
-	ExecuteGovernanceRuleParams *ExecuteGovernanceRuleParams
+// GovernanceRulesClientBeginDeleteOptions contains the optional parameters for the GovernanceRulesClient.BeginDelete method.
+type GovernanceRulesClientBeginDeleteOptions struct {
 	// Resumes the LRO from the provided token.
 	ResumeToken string
 }
 
-// GovernanceRulesClientBeginRuleIDExecuteSingleSubscriptionOptions contains the optional parameters for the GovernanceRulesClient.BeginRuleIDExecuteSingleSubscription
-// method.
-type GovernanceRulesClientBeginRuleIDExecuteSingleSubscriptionOptions struct {
-	// GovernanceRule over a subscription scope
+// GovernanceRulesClientBeginExecuteOptions contains the optional parameters for the GovernanceRulesClient.BeginExecute method.
+type GovernanceRulesClientBeginExecuteOptions struct {
+	// Execute governance rule over a given scope
 	ExecuteGovernanceRuleParams *ExecuteGovernanceRuleParams
 	// Resumes the LRO from the provided token.
 	ResumeToken string
@@ -4249,13 +4312,19 @@ type GovernanceRulesClientCreateOrUpdateOptions struct {
 	// placeholder for future optional parameters
 }
 
-// GovernanceRulesClientDeleteOptions contains the optional parameters for the GovernanceRulesClient.Delete method.
-type GovernanceRulesClientDeleteOptions struct {
+// GovernanceRulesClientGetOptions contains the optional parameters for the GovernanceRulesClient.Get method.
+type GovernanceRulesClientGetOptions struct {
 	// placeholder for future optional parameters
 }
 
-// GovernanceRulesClientGetOptions contains the optional parameters for the GovernanceRulesClient.Get method.
-type GovernanceRulesClientGetOptions struct {
+// GovernanceRulesClientListOptions contains the optional parameters for the GovernanceRulesClient.List method.
+type GovernanceRulesClientListOptions struct {
+	// placeholder for future optional parameters
+}
+
+// GovernanceRulesClientOperationResultsOptions contains the optional parameters for the GovernanceRulesClient.OperationResults
+// method.
+type GovernanceRulesClientOperationResultsOptions struct {
 	// placeholder for future optional parameters
 }
 
@@ -5702,6 +5771,12 @@ type OperationList struct {
 	NextLink *string `json:"nextLink,omitempty" azure:"ro"`
 }
 
+// OperationResultAutoGenerated - Long run operation status of governance rule over a given scope
+type OperationResultAutoGenerated struct {
+	// READ-ONLY; The status of the long run operation result of governance rule
+	Status *OperationResult `json:"status,omitempty" azure:"ro"`
+}
+
 // OperationsClientListOptions contains the optional parameters for the OperationsClient.List method.
 type OperationsClientListOptions struct {
 	// placeholder for future optional parameters
@@ -7091,13 +7166,6 @@ type SubAssessmentsClientListAllOptions struct {
 // SubAssessmentsClientListOptions contains the optional parameters for the SubAssessmentsClient.List method.
 type SubAssessmentsClientListOptions struct {
 	// placeholder for future optional parameters
-}
-
-// SubscriptionGovernanceRulesExecuteStatusClientBeginGetOptions contains the optional parameters for the SubscriptionGovernanceRulesExecuteStatusClient.BeginGet
-// method.
-type SubscriptionGovernanceRulesExecuteStatusClientBeginGetOptions struct {
-	// Resumes the LRO from the provided token.
-	ResumeToken string
 }
 
 type SuppressionAlertsScope struct {